refactor: rename wallet management components as account management
