+++ packages/features/src/account-management/components/details-dropdown.tsx
